refactor: Layer-first to Feature-first
