+++ embed/lib/src/common/embedder.dart
@@ -6,6 +6,11 @@ import 'package:embed_annotation/embed_annotation.dart';
 
 abstract class Embedder<E extends Embed> {
   const Embedder(this.config);
+
   final E config;
-  FutureOr<String> getEmbeddingOf(File content, TopLevelVariableElement element);
+
+  FutureOr<String> getEmbeddingOf(
+    File content,
+    TopLevelVariableElement element,
+  );
 }
