+++ embed/lib/src/literal/literal_embedder.dart
@@ -3,9 +3,9 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:analyzer/dart/element/element.dart';
-import 'package:embed/src/common/pattern_matching.dart';
-import 'package:embed/src/common/type_constraints.dart';
-import 'package:embed/src/embedders/embedder.dart';
+import 'package:embed/src/literal/pattern_matching.dart';
+import 'package:embed/src/literal/type_constraints.dart';
+import 'package:embed/src/common/embedder.dart';
 import 'package:embed_annotation/embed_annotation.dart';
 import 'package:path/path.dart' as p;
 import 'package:toml/toml.dart';
