+++ embed/lib/src/literal/type_constraints.dart
