+++ embed/lib/src/str/embed_str_generator.dart
@@ -1,6 +1,6 @@
-import 'package:embed/src/embedders/embedder.dart';
-import 'package:embed/src/embedders/string_embedder.dart';
-import 'package:embed/src/generators/embed_generator.dart';
+import 'package:embed/src/common/embedder.dart';
+import 'package:embed/src/str/string_embedder.dart';
+import 'package:embed/src/common/embed_generator.dart';
 import 'package:embed_annotation/embed_annotation.dart';
 import 'package:source_gen/source_gen.dart';
