+++ embed/lib/src/utils.dart
