+++ embed/test/generators/str/embed_str_test.dart
@@ -1,4 +1,4 @@
-import 'package:embed/src/generators/embed_str_generator.dart';
+import 'package:embed/src/str/embed_str_generator.dart';
 import 'package:source_gen_test/source_gen_test.dart';
 
 Future<void> main() async {
